refactor(search applications): rename engine connect folder
