+++ x-pack/plugins/enterprise_search/public/applications/applications/components/search_application/connect/search_application_documentation.tsx
